requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-experimental>=0.4.0
neo4j>=5.0.0
python-dotenv>=1.0.0

# Optional: Web Interface
streamlit>=1.28.0
streamlit-agraph>=0.0.45

# Development
openai>=1.0.0
